Optimization: Moving cursor does not redraw Grid
The grid component referenced selected_x and selected_y. This led
to the grid being rendered each time either one of them changes,
ie. when the cursor is moved. When the canvas is large this leads
to frustrating lag.

This commit extracts the row and column numbers, as well as the
cursor to separate components outside the grid component, leading
to moving around being fast :)